NOISSUE: clean integr tests
